# sidecar/openapi.yaml
...
openapi: 3.0.3
info:
  title: Clinical Co-Pilot Sidecar API
  description: >
    Preparation pipeline, fact store, deterministic citation gate, document
    ingestion (attach_and_extract), hybrid guideline retrieval, and streaming
    chat beside an untouched OpenEMR. Every response carries x-correlation-id;
    a full trace reconstructs from that ID alone (G4).
  version: "2.0.0"
  license:
    name: GPL-3.0
servers:
  - url: http://localhost:8080
    description: local dev
  - url: https://{railway-app}.up.railway.app
    description: Railway deploy (public demo)
...
paths:
...
  /api/chat/{patientId}:
    post:
      tags: [chat]
      summary: One chat turn, streamed as SSE.
      description: >
        Guards answer as plain JSON before the stream opens (400 invalid message,
        404 patient, 429 budget). After the stream opens, events are
        seed | status | delta | citation | tool_use | tool_result | done | error.
        Week 2 (E.9): guideline-shaped questions take the evidence lane — a
        `status` event ("checking practice protocols…") then a critic-verified
        answer whose citations are guideline_evidence; graph failures fall back
        to the Week 1 loop. done carries { conversation_id, citations,
        unverified_count, tools_used, prescriptive_flag_count }.
      parameters: [{ name: patientId, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string, maxLength: 2000 }
                conversation_id: { type: string }
                viewing_image_id: { type: string }
      responses:
        "200": { description: "text/event-stream of chat events" }
        "400": { description: invalid_message }
        "404": { description: patient_not_found }
        "429": { description: llm_budget_exceeded }
        "503": { description: store_not_configured }
    get:
      tags: [chat]
      summary: Replay a conversation transcript (persistence across reloads).
      parameters:
        - { name: patientId, in: path, required: true, schema: { type: string } }
        - { name: conversation_id, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: "{ messages: StoredChatMessage[] }" }
        "503": { description: store_not_configured }
...